Unify user and page create/update forms
