+++ crates/brace-web-auth/src/lib/form/mod.rs
@@ -1,3 +1,2 @@
-pub mod create;
 pub mod login;
-pub mod update;
+pub mod user;
