+++ crates/brace-web-page/src/lib/form/mod.rs
@@ -1,2 +1,1 @@
-pub mod create;
-pub mod update;
+pub mod page;
