+++ crates/brace-web-page/src/lib/route/web/create.rs
@@ -9,7 +9,7 @@ use brace_web_form::Form;
 use futures::future::{err, Either, Future};
 use serde_json::json;
 
-use crate::form::create::CreateForm;
+use crate::form::page::PageForm;
 use crate::model::{Page, PageWithPath};
 
 pub fn get(
@@ -46,7 +46,7 @@ fn render(
     pages: Vec<PageWithPath>,
     renderer: &Renderer,
 ) -> impl Future<Item = HttpResponse, Error = Error> {
-    let form = Form::build(CreateForm, (), pages).unwrap();
+    let form = Form::build(PageForm, Page::default(), pages).unwrap();
     let template = Template::new(
         "form-layout",
         json!({
